requirements.txt
# ── Web Framework ─────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9       # file uploads

# ── Database ──────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1               # migrations DB
asyncpg==0.29.0               # PostgreSQL async driver
psycopg2-binary==2.9.9

# ── Task Queue ────────────────────────────────────────
celery==5.4.0
redis==5.0.4

# ── LLM & Agents ──────────────────────────────────────
langchain==0.2.5
langchain-openai==0.1.8
langchain-anthropic==0.1.15
langchain-community==0.2.5
langgraph==0.1.14
langsmith==0.1.75
******************************
PyJWT==2.8.1
passlib[bcrypt]==1.7.4
email-validator==2.1.1
python-jose[cryptography]==3.3.0

# ── Vector Store ──────────────────────────────────────
chromadb==0.5.0
sentence-transformers==3.0.1  # embeddings locaux (gratuit)

# ── PDF / PPTX Parsing ────────────────────────────────
pymupdf==1.24.5               # extraction PDF
python-pptx==0.6.23           # extraction PPTX

# ── Browser Agent ─────────────────────────────────────
browser-use==0.1.16
playwright==1.44.0

# ── Web Search ────────────────────────────────────────
tavily-python==0.3.3

# ── Auth & Security ───────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Config ────────────────────────────────────────────
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ── Utils ─────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1

# ── Tests ─────────────────────────────────────────────
pytest==8.2.1
pytest-asyncio==0.23.7
httpx==0.27.0
